+++ Sources/Extensions/String+Version.swift
@@ -1,3 +1,5 @@
+// periphery:ignore:all
+
 import Foundation
 
 public typealias VersionString = String