+++ Sources/Shared/SwiftVersion.swift
@@ -2,7 +2,7 @@ import Extensions
 import Foundation
 
 public struct SwiftVersion {
-    static let minimumVersion = "5.10"
+    static let minimumVersion = "6.1.2"
 
     public let version: VersionString
     public let fullVersion: String