+++ Sources/SourceGraph/Mutators/ExternalOverrideRetainer.swift
@@ -8,27 +8,20 @@ import Shared
 /// as the external implementation may call the overridden declaration.
 final class ExternalOverrideRetainer: SourceGraphMutator {
     private let graph: SourceGraph
-    private let isSwift60FixEnabled: Bool
 
-    required init(graph: SourceGraph, configuration _: Configuration, swiftVersion: SwiftVersion) {
+    required init(graph: SourceGraph, configuration _: Configuration, swiftVersion _: SwiftVersion) {
         self.graph = graph
-        isSwift60FixEnabled = swiftVersion.version.isVersion(greaterThanOrEqualTo: "6.0") &&
-            swiftVersion.version.isVersion(lessThan: "6.1")
     }
 
     func mutate() {
         for decl in graph.declarations(ofKinds: Declaration.Kind.overrideKinds) {
             guard decl.isOverride else { continue }
 
-            var didIdentifyRelatedRef = false
-
             for relatedRef in decl.related {
                 if relatedRef.declarationKind == decl.kind,
                    relatedRef.name == decl.name,
                    relatedRef.location == decl.location
                 {
-                    didIdentifyRelatedRef = true
-
                     if graph.declaration(withUsr: relatedRef.usr) == nil {
                         // The related decl is external.
                         graph.markRetained(decl)
@@ -36,11 +29,6 @@ final class ExternalOverrideRetainer: SourceGraphMutator {
 
                     break
                 }
-            }
-
-            // https://github.com/swiftlang/swift/issues/76628
-            if !didIdentifyRelatedRef, isSwift60FixEnabled {
-                graph.markRetained(decl)
             }
         }
     }